+++ hera_sim/tests/test_interpolators.py
@@ -5,8 +5,8 @@ import os
 import glob
 import tempfile
 
-from scipy.interpolate import RectBivariateSpline
-from hera_sim.interpolators import Tsky
+from scipy.interpolate import RectBivariateSpline, interp1d
+from hera_sim.interpolators import Tsky, Beam, Bandpass
 from nose.tools import raises
 
 def test_interpolator():
@@ -66,3 +66,56 @@ def test_bad_npz():
     # now check that it catches using a bad polarization
     np.savez(temp_dir+'/ok_file', lsts=lsts, freqs=freqs, tsky=ok_tsky, meta=meta)
     tsky = Tsky(temp_dir+'/ok_file.npz', pol='yy')
+
+def test_FreqInterpolators():
+    # make a temporary directory
+    temp_dir = tempfile.mkdtemp()
+
+    # make some mock data
+    freqs = np.linspace(0.1,0.2,100)
+    values = np.arange(freqs.size)
+
+    # get a polyfit
+    polyfit = np.polyfit(freqs, values, deg=1)
+
+    # save some files
+    np.save(os.path.join(temp_dir, 'polyfit'), polyfit)
+    np.savez(os.path.join(temp_dir, 'beam'), freqs=freqs, beam=values)
+
+    # check that things work as expected for a poly1d interpolator
+    interp = Bandpass(os.path.join(temp_dir, 'polyfit.npy'), interpolator='poly1d')
+    assert isinstance(interp._interpolator, np.poly1d)
+    assert interp(freqs).size == freqs.size
+    
+    # now do the same for a interp1d interpolator
+    interp = Beam(os.path.join(temp_dir, 'beam.npz'), interpolator='interp1d')
+    assert isinstance(interp._interpolator, interp1d)
+    assert interp(freqs).size == freqs.size
+
+@raises(AssertionError)
+def test_bad_params():
+    # make a temporary directory
+    temp_dir = tempfile.mkdtemp()
+
+    # make some mock data
+    freqs = np.linspace(0.1,0.2,100)
+    values = np.arange(freqs.size)
+
+    # make some bad files
+    np.savez(os.path.join(temp_dir, 'no_freqs'), beam=values)
+    np.savez(os.path.join(temp_dir, 'no_values'), freqs=freqs)
+    np.save(os.path.join(temp_dir, 'some_npy'), values)
+
+    # now try to make Beam objects with bad files/parameters
+    # bad interpolator
+    interp = Beam(os.path.join(temp_dir, 'some_npy.npy'), interpolator='something')
+    # bad datafile extension v1
+    interp = Beam(os.path.join(temp_dir, 'some_npy.npy'), interpolator='interp1d')
+    # bad datafile extension v2
+    interp = Beam(os.path.join(temp_dir, 'no_freqs.npz'), interpolator='poly1d')
+    # bad keys
+    interp = Beam(os.path.join(temp_dir, 'no_freqs.npz'), interpolator='interp1d')
+    interp = Beam(os.path.join(temp_dir, 'no_values.npz'), interpolator='interp1d')
+    # nonexistent file
+    interp = Beam(os.path.join(temp_dir, 'not_a_file.npz'), interpolator='interp1d')
+